third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animation-effects/phases-and-states-expected.txt

This is a testharness.js-based test.
PASS Phase calculation for a simple animation effect
PASS Phase calculation for an animation effect with a positive start delay
FAIL Phase calculation for an animation effect with a negative start delay assert_equals: Animation effect is in before phase when current time is -1ms (progress is null with 'none' fill mode) expected (object) null but got (number) 0
PASS Phase calculation for an animation effect with a positive end delay
PASS Phase calculation for an animation effect with a negative end delay lesser in magnitude than the active duration
PASS Phase calculation for an animation effect with a negative end delay equal in magnitude to the active duration
FAIL Phase calculation for an animation effect with a negative end delay greater in magnitude than the active duration assert_not_equals: Animation effect is in before phase when current time is -1ms (progress is non-null with appropriate fill mode) got disallowed value null
PASS Phase calculation for an animation effect with a positive start delay and a negative end delay lesser in magnitude than the active duration
FAIL Phase calculation for an animation effect with a negative start delay and a negative end delay equal in magnitude to the active duration assert_not_equals: Animation effect is in before phase when current time is -1ms (progress is non-null with appropriate fill mode) got disallowed value null
FAIL Phase calculation for an animation effect with a negative start delay and a negative end delay equal greater in magnitude than the active duration assert_not_equals: Animation effect is in before phase when current time is -2ms (progress is non-null with appropriate fill mode) got disallowed value null
FAIL Phase calculation for a simple animation effect with negative playback rate assert_not_equals: Animation effect is in active phase when current time is 1ms got disallowed value null
Harness: the test ran to completion.
